Add files via upload
This is the first website that I developed! I got the template from [Dicoding](https://www.dicoding.com/), even though I found so many new things when developing it, but I found it fun to develop a website like this!

This is just a simple elementary web calculator with only +/-or negative number operations in it and some calculation history.