Create code_generator.ipynb
This is "code generator" code, the purpose is to create the next code after the one before, you can apply this code if you have/will have a collecting bot on discord such as "Karuta", "Tofu", or anything that similar to it or even something new that needed this code. Just add few lines to store your last_code or make few changes on this code by your own!